Migrated graceful tests
